+++ lib/folio/engine.rb
@@ -48,6 +48,7 @@ module Folio
     config.folio_direct_s3_upload_class_names = %w[
       Folio::File
       Folio::PrivateAttachment
+      Folio::SessionAttachment::Base
     ]
 
     config.folio_direct_s3_upload_allow_for_users = false